Moved InMemoryDataManager as a separate test only DataManager class
